requirements.txt
streamlit
pypdf
python-docx
openai